Moved play code into new TTGGame class
